project_generate: Removed all hard coded paths and added optional input arguments instead.
Added options to control where outputs are read from and written to
these include:
--rootdir = Can be used to specify the location of a valid configure
script. This allows the program to be run from any directory as opposed
to the previous hard coded path. If none is specified the generator will
search some common paths in order to find it.
--projdir = This allows for specifying where the output project files
are created. This allows for out of tree creation locations etc. If not
specified it will create a default 'SMP' directory in the ROOT path.
--prefix = This specifies where the created projects will output there
final build results. This is also the location that they search for
dependencies (This is not complete at this stage and will be finalized
in next commit).